backend/requirements.txt
flask~=2.2.5
numpy~=1.24.3
requests~=2.31.0
pydecision~=4.5.8
python-dotenv~=1.0.1